# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Question Generator API
  version: 1.0.0
  description: API documentation for the AI-powered Question Generation Tool
servers:
  - url: http://localhost:5000/api/v1
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User registered
        '400':
          description: Bad request
  /auth/login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '400':
          description: Bad request
  /questions/generate:
    post:
      summary: Generate questions using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                chapter:
                  type: string
                difficulty:
                  type: string
                type:
                  type: string
                count:
                  type: integer
                concepts:
                  type: array
                  items:
                    type: string
                exclude_patterns:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Questions generated
  /questions/search:
    get:
      summary: Search questions
      parameters:
        - in: query
          name: subject
          schema:
            type: string
        - in: query
          name: chapter
          schema:
            type: string
        - in: query
          name: difficulty
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        '200':
          description: Questions found
  /users/profile:
    get:
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
    put:
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                subscription_tier:
                  type: string
      responses:
        '200':
          description: User profile updated
  /api-keys/generate:
    post:
      summary: Generate API key
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions:
                  type: object
                rate_limits:
                  type: object
      responses:
        '201':
          description: API key generated
  /api-keys:
    get:
      summary: List API keys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of API keys
  /api-keys/{id}:
    delete:
      summary: Revoke API key
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key revoked
  /analytics/usage/statistics:
    get:
      summary: Get usage statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usage statistics
  /analytics/performance:
    get:
      summary: Get performance analytics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Performance analytics
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service health
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 