# src/routes/v1/auth.routes.ts
import { Router } from 'express';
import { registerUser, loginUser, generateRefreshToken, verifyRefreshToken } from '../../services/auth.service';
import jwt from 'jsonwebtoken';
import { z } from 'zod';
import { validate } from '../../middleware/validate';

const router = Router();

const authSchema = z.object({
  email: z.string().email(),
  password: z.string().min(6),
});

router.post('/register', validate(authSchema), async (req, res) => {
  try {
    const { email, password } = req.body;
    const user = await registerUser(email, password);
    res.status(201).json({ user });
  } catch (err: any) {
    res.status(400).json({ error: err.message });
  }
});

router.post('/login', validate(authSchema), async (req, res) => {
  try {
    const { email, password } = req.body;
    const { token, user } = await loginUser(email, password);
    res.json({ token, user });
  } catch (err: any) {
    res.status(400).json({ error: err.message });
  }
});

router.post('/refresh', (req, res) => {
  const { refreshToken } = req.body;
  if (!refreshToken) return res.status(400).json({ error: 'Refresh token required' });
  const payload = verifyRefreshToken(refreshToken);
  if (!payload) return res.status(401).json({ error: 'Invalid or expired refresh token' });
  const token = jwt.sign({ userId: (payload as any).userId, email: (payload as any).email }, process.env.JWT_SECRET || 'changeme', { expiresIn: '1h' });
  res.json({ token });
});

export default router; 